Allow addons to hide loading tips
Addons are able to customize the number of contents of loading
screen tips via scripts and resource XML files, but they are not
able to remove the tips altogether because empty strings are not
allowed as resource strings and the engine always prints out a
hardcoded tip number followed by a colon (:) anyway.

With this minor change,  `loadscreen.script` is able to return `0` for
`get_tip_number` which will hide the loading tip altogether.